+++ tests/integration/test_models/test_full_battery_models/test_lithium_ion/base_lithium_ion_tests.py
@@ -3,6 +3,7 @@
 #
 import pybamm
 import tests
+
 import numpy as np